Keeps the SyncEngine alive when the Wiredash widget updates
The SyncEngine (or better their jobs) depends on the WiredashApi. The WiredashApi depends on the Wiredash widget. When the Wiredash widget changes -because the secret or the projectId have might have changed - the WiredashApi gets recreated and so does the SyncEngine. Unfortunately, this cancels the inital sync where the engine waits 5s before firing the first jobs.

This fix makes sure the SyncEngine does not get recreated. It uses the same instance as long as Wiredash lives.

Why wasn't this caught with tests? Well it was tested. But in tests there is no dependency from the MockWireadshApi to the Wiredash Widget, thus not causing a rebuild. Sigh...