User choses to submit their email or not
When a prefilled email address was set, it was sent to be backend regardless if the user deleted it in the input field or not. This is now fixed and transparent. The email is now submitted as the user sees it in the email step. It might be prefilled or not. When the field is empty nothing is sent