Make trackEvent() not wait for submission
Previously, the trackEvent() future completed when a request with that event was sent to the server (ignoring errors). This interfered with the debounce of 5s after app start, causing trackEvents at app start to complete only after 5s.

Now trackEvent() completes when the event is saved to disk. The submission later is handled internally by the SDK.

Fixes https://github.com/wiredashio/wiredash-sdk/issues/368